+++ include/sorter.h
@@ -44,7 +44,7 @@ namespace internal
  */
 template <typename T>
 typename std::enable_if<std::is_same<T, int>::value>::type
-sorter(T*& data, uint64_t size);
+sorter(T*& data, uint32_t size);
 
 /**
  * Integer version (key, ptr):
@@ -59,27 +59,27 @@ sorter(T*& data, uint64_t size);
  */
 template <typename T>
 typename std::enable_if<std::is_same<T, int>::value>::type
-sorter_key(T*& data, int *&ptr, uint64_t size);
+sorter_key(T*& data, int *&ptr, uint32_t size);
 
 template <typename T>
 typename std::enable_if<std::is_same<T, float>::value>::type
-sorter_key(T*& data, int *&ptr, uint64_t size);
+sorter_key(T*& data, int *&ptr, uint32_t size);
 
 template <typename T>
 typename std::enable_if<std::is_same<T, double>::value>::type
-sorter_key(T*& data, int *&ptr, uint64_t size);
+sorter_key(T*& data, int *&ptr, uint32_t size);
 
 template <typename T>
 typename std::enable_if<std::is_same<T, int>::value>::type
-sorter_key(T*& data, long *&ptr, uint64_t size);
+sorter_key(T*& data, long *&ptr, uint32_t size);
 
 template <typename T>
 typename std::enable_if<std::is_same<T, float>::value>::type
-sorter_key(T*& data, long *&ptr, uint64_t size);
+sorter_key(T*& data, long *&ptr, uint32_t size);
 
 template <typename T>
 typename std::enable_if<std::is_same<T, double>::value>::type
-sorter_key(T*& data, long *&ptr, uint64_t size);
+sorter_key(T*& data, long *&ptr, uint32_t size);
 
 
 
@@ -95,7 +95,7 @@ sorter_key(T*& data, long *&ptr, uint64_t size);
  */
 template <typename T>
 typename std::enable_if<std::is_same<T, float>::value>::type
-sorter(T*& data, uint64_t size);
+sorter(T*& data, uint32_t size);
 
 /**
  * Double version:
@@ -109,7 +109,7 @@ sorter(T*& data, uint64_t size);
  */
 template <typename T>
 typename std::enable_if<std::is_same<T, double>::value>::type
-sorter(T*& data, uint64_t size);
+sorter(T*& data, uint32_t size);
 
 
 } // end namespace internal
